Move common update lines component to new file
