+++ server/service/src/invoice_line/outbound_shipment_line/update/mod.rs
@@ -1,38 +1,27 @@
 use crate::{
-    invoice_line::{query::get_invoice_line, ShipmentTaxUpdate},
+    invoice_line::{
+        common_insert_update::{generate, validate, UpdateInvoiceLine, UpdateInvoiceLineError},
+        query::get_invoice_line,
+    },
     service_provider::ServiceContext,
-    WithDBError,
 };
-use repository::{
-    InvoiceLine, InvoiceLineRow, InvoiceLineRowRepository, RepositoryError, StockLineRow,
-    StockLineRowRepository,
-};
+use repository::{InvoiceLine, InvoiceLineRowRepository, InvoiceRowType, StockLineRowRepository};
 
-mod generate;
-mod validate;
-
-use generate::generate;
-use validate::validate;
-#[derive(Clone, Debug, PartialEq, Default)]
-pub struct UpdateOutboundShipmentLine {
-    pub id: String,
-    pub item_id: Option<String>,
-    pub stock_line_id: Option<String>,
-    pub number_of_packs: Option<f64>,
-    pub total_before_tax: Option<f64>,
-    pub tax: Option<ShipmentTaxUpdate>,
-}
-
-type OutError = UpdateOutboundShipmentLineError;
+type OutError = UpdateInvoiceLineError;
 
 pub fn update_outbound_shipment_line(
     ctx: &ServiceContext,
-    input: UpdateOutboundShipmentLine,
+    input: UpdateInvoiceLine,
 ) -> Result<InvoiceLine, OutError> {
     let updated_line = ctx
         .connection
         .transaction_sync(|connection| {
-            let (line, item, batch_pair, invoice) = validate(&input, &ctx.store_id, &connection)?;
+            let (line, item, batch_pair, invoice) = validate(
+                &input,
+                &ctx.store_id,
+                &connection,
+                InvoiceRowType::OutboundShipment,
+            )?;
 
             let (update_line, batch_pair) = generate(input, line, item, batch_pair, invoice)?;
             InvoiceLineRowRepository::new(&connection).upsert_one(&update_line)?;
@@ -49,82 +38,6 @@ pub fn update_outbound_shipment_line(
         })
         .map_err(|error| error.to_inner_error())?;
     Ok(updated_line)
-}
-/// During outbound shipment line update, stock line may change thus validation and updates need to apply to both batches
-pub struct BatchPair {
-    /// Main batch to be updated
-    pub main_batch: StockLineRow,
-    /// Optional previous batch (if batch was changed)
-    pub previous_batch_option: Option<StockLineRow>,
-}
-
-impl BatchPair {
-    /// Calculate reduction amount to apply to main batch
-    pub fn get_main_batch_reduction(
-        &self,
-        input: &UpdateOutboundShipmentLine,
-        existing_line: &InvoiceLineRow,
-    ) -> f64 {
-        // Previous batch exists, this mean new batch was requested means:
-        // - reduction should be number of packs from input (or existing line if number of pack is missing in input)
-        if self.previous_batch_option.is_some() {
-            input
-                .number_of_packs
-                .unwrap_or(existing_line.number_of_packs)
-        } else {
-            // Previous batch does not exists, this mean updating existing batch, thus:
-            // - reduction is the difference between input and existing line number of packs
-            if let Some(number_of_packs) = &input.number_of_packs {
-                *number_of_packs - existing_line.number_of_packs
-            } else {
-                // No changes in input, no reduction
-                0.0
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum UpdateOutboundShipmentLineError {
-    LineDoesNotExist,
-    DatabaseError(RepositoryError),
-    InvoiceDoesNotExist,
-    NotAnOutboundShipment,
-    NotThisStoreInvoice,
-    NotThisInvoiceLine(String),
-    CannotEditFinalised,
-    ItemNotFound,
-    StockLineNotFound,
-    NumberOfPacksBelowOne,
-    ItemDoesNotMatchStockLine,
-    LocationIsOnHold,
-    LocationNotFound,
-    LineDoesNotReferenceStockLine,
-    BatchIsOnHold,
-    UpdatedLineDoesNotExist,
-    StockLineAlreadyExistsInInvoice(String),
-    ReductionBelowZero {
-        stock_line_id: String,
-        line_id: String,
-    },
-}
-
-impl From<RepositoryError> for UpdateOutboundShipmentLineError {
-    fn from(error: RepositoryError) -> Self {
-        UpdateOutboundShipmentLineError::DatabaseError(error)
-    }
-}
-
-impl<ERR> From<WithDBError<ERR>> for UpdateOutboundShipmentLineError
-where
-    ERR: Into<UpdateOutboundShipmentLineError>,
-{
-    fn from(result: WithDBError<ERR>) -> Self {
-        match result {
-            WithDBError::DatabaseError(error) => error.into(),
-            WithDBError::Error(error) => error.into(),
-        }
-    }
 }
 
 #[cfg(test)]
@@ -145,8 +58,8 @@ mod test {
 
     use crate::{
         invoice::outbound_shipment::{UpdateOutboundShipment, UpdateOutboundShipmentStatus},
-        invoice_line::outbound_shipment_line::{
-            update::UpdateOutboundShipmentLine, UpdateOutboundShipmentLineError as ServiceError,
+        invoice_line::common_insert_update::{
+            UpdateInvoiceLine, UpdateInvoiceLineError as ServiceError,
         },
         service_provider::ServiceProvider,
     };
@@ -169,7 +82,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = "invalid".to_string();
                 }),
             ),
@@ -180,7 +93,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(10.0);
                 }),
@@ -192,7 +105,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_inbound_shipment_a_invoice_lines()[0].id.clone();
                 }),
             ),
@@ -204,7 +117,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_b_invoice_lines()[0].id.clone();
                 }),
             ),
@@ -215,7 +128,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_no_stock_line()[0].id.clone();
                 }),
             ),
@@ -227,7 +140,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.item_id = Some("invalid".to_string());
                 }),
@@ -239,7 +152,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some("invalid".to_string());
                 }),
@@ -251,7 +164,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(0.0);
                 }),
@@ -263,7 +176,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.item_id = Some(mock_item_b().id.clone());
                     r.stock_line_id = Some(mock_stock_line_a().id.clone());
@@ -276,7 +189,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.item_id = Some(mock_stock_line_location_is_on_hold()[0].item_id.clone());
                     r.stock_line_id = Some(mock_stock_line_location_is_on_hold()[0].id.clone());
@@ -289,7 +202,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some(mock_stock_line_on_hold()[0].id.clone());
                     r.item_id = Some(mock_stock_line_on_hold()[0].item_id.clone());
@@ -302,7 +215,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(100.0);
                 }),
@@ -320,7 +233,7 @@ mod test {
         assert_eq!(
             service.update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some(mock_item_b_lines()[0].id.clone());
                 }),
@@ -375,7 +288,7 @@ mod test {
         service
             .update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_c_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(2.0);
                     r.total_before_tax = Some(18.00);
@@ -435,7 +348,7 @@ mod test {
         service
             .update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_c_invoice_lines()[0].id.clone();
                     r.stock_line_id = Some(mock_stock_line_b().id.clone());
                     r.number_of_packs = Some(2.0);
@@ -477,7 +390,7 @@ mod test {
         service
             .update_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut UpdateOutboundShipmentLine| {
+                inline_init(|r: &mut UpdateInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.number_of_packs = Some(15.0);
                     r.total_before_tax = Some(10.99);
